Начальный вариант Dockerfile

+++ src/main/java/ru/yandex/practicum/catsgram/model/Post.java
@@ -7,7 +7,7 @@ import java.time.Instant;
 @Data
 public class Post {
     private Long id;
-    private long authorId;
+    private Integer authorId;
     private String description;
     private Instant postDate;
 }
